+++ functions/common.js
